Added go.mod to create a module. eliminating relative ugly paths! also changed names of functions, imports, and files for ease of readability
