+++ backend/go.mod
@@ -1,0 +1,18 @@
+module github.com/omaralaniz/omar
+
+go 1.13
+
+require (
+	github.com/DataDog/zstd v1.4.4 // indirect
+	github.com/gin-gonic/gin v1.5.0
+	github.com/go-stack/stack v1.8.0 // indirect
+	github.com/golang/snappy v0.0.1 // indirect
+	github.com/joho/godotenv v1.3.0
+	github.com/pkg/errors v0.8.1 // indirect
+	github.com/xdg/scram v0.0.0-20180814205039-7eeb5667e42c // indirect
+	github.com/xdg/stringprep v1.0.0 // indirect
+	go.mongodb.org/mongo-driver v1.2.0
+	golang.org/x/crypto v0.0.0-20191206172530-e9b2fee46413 // indirect
+	golang.org/x/sync v0.0.0-20190911185100-cd5d95a43a6e // indirect
+	golang.org/x/text v0.3.2 // indirect
+)
