--- backend/handlers/projects_get.go
@@ -1,15 +1,0 @@
-package handlers
-
-import (
-	"net/http"
-
-	"../repository"
-	"github.com/gin-gonic/gin"
-)
-
-func GetAll(projects *repository.Projects) gin.HandlerFunc {
-	return func(c *gin.Context) {
-		results := projects.GetAll()
-		c.JSON(http.StatusOK, results)
-	}
-}
