+++ backend/main.go
@@ -1,18 +1,18 @@
 package main
 
 import (
-	"./handlers"
-	"./repository"
+	handler "github.com/omaralaniz/omar/handlers"
+	repo "github.com/omaralaniz/omar/repository"
 	"github.com/gin-gonic/gin"
 )
 
 func main() {
 	r := gin.Default()
-	projects := repository.New()
+	projects := repo.New()
 
 	api := r.Group("/api")
-	api.GET("/ping", handlers.Ping())
-	api.GET("/projects", handlers.GetAll(projects))
+	api.GET("/ping", handler.Ping())
+	api.GET("/projects", handler.GetAll(projects))
 
 	err := r.Run("0.0.0.0:8080")
